Remove redundant and unnecessary method "headers_to_array":
As indicated in the discussion on #911, the "headers_to_array"
method, which was originally available through
Dancer2::Core::Role::Headers, was copied over verbatim. In the
discussion, @veryrusty mentioned it is indeed redundant since it
is only used by a response. A request does not need to flatten its
headers.

Thus the method was removed.